Add files via upload
This Movie Recommendation System uses the TMDB 5000 Movies Dataset to provide personalized movie suggestions. By leveraging content-based filtering and cosine similarity, it analyzes movie attributes like genres, cast, crew, and overviews to generate recommendations. The system achieves an 80% accuracy in recommendations, with a diversity score of 0.75 and a personalization score of 0.65, ensuring tailored and varied suggestions. Evaluated using Hit Rate@5 (0.85) and F1-score (0.72), it delivers high-quality, user-centric recommendations. Integrated with Streamlit and the TMDB API, it offers a seamless and visually appealing user experience. This project showcases the power of data-driven techniques in creating engaging and personalized entertainment solutions.